Move file read and write methods from BaseHandler to FileStorage
