+++ src/pricecypher/storage/file_storage.py
@@ -2,13 +2,16 @@ from __future__ import annotations
 
 import logging
 import os
+import pandas as pd
+import pyarrow as pa
+import pyarrow.parquet as pq
 import re
+import smart_open
+
 from abc import ABC
 from contextlib import contextmanager
 from pathlib import Path
 from typing import Type, Union, Optional
-
-import smart_open
 
 from pricecypher.dataclasses import HandlerSettings
 from pricecypher.dataclasses.settings.remote_storage_settings import RemoteStorageSettings
@@ -133,3 +136,75 @@ class FileStorage(ABC):
             path_remote_prefix=settings.path_remote_out_prefix,
             remote_storage_settings=settings.remote_storage_settings,
         )
+
+    def read_file(self, path_in: str) -> str:
+        """
+        Read a string file from path_in, using FileStorage.
+        :param path_in: the path to read from.
+        :return: the file contents.
+        """
+        with self.load(path_in, mode='r') as f:
+            return f.read()
+
+    def write_file(self, path_out: str, string: str) -> str:
+        """
+        Write a string to a file, using FileStorage.
+        :param path_out: the path to write the file to.
+        :param string: the string to write to the file.
+        :return: the remote storage path.
+        """
+        self._create_folders(path_out)
+        with self.save(path_out, mode='w') as f:
+            f.write(string)
+        return self.get_path_remote(path_out)
+
+    def read_df(self, path_in: str) -> pd.DataFrame:
+        """
+        Read a DataFrame pickle from path_in, using FileStorage.
+        :param path_in: the path to read from.
+        :return: the DataFrame.
+        """
+        with self.load(path_in, mode='rb') as f:
+            return pd.read_pickle(f)
+
+    def write_df(self, path_out: str, df: pd.DataFrame) -> str:
+        """
+        Write a DataFrame to a pickle, using FileStorage.
+        :param path_out: the path to write the pickle to.
+        :param df: the DataFrame to store.
+        :return: the remote storage path.
+        """
+        self._create_folders(path_out)
+        with self.save(path_out, mode='wb') as f:
+            pd.to_pickle(df, f)
+        return self.get_path_remote(path_out)
+
+    def read_parquet(self, path_in: str) -> pa.Table:
+        """
+        Read a parquet file from path_in, using FileStorage.
+        :param path_in: the path to read from.
+        :return: the DataFrame.
+        """
+        with self.load(path_in, mode='rb') as f:
+            return pq.read_table(f)
+
+    def write_parquet(self, path_out: str, table: pa.Table) -> str:
+        """
+        Write a parquet file, using FileStorage.
+        :param path_out: the path to write the parquet file to.
+        :param table: the Table to store.
+        :return: the remote storage path.
+        """
+        self._create_folders(path_out)
+        with self.save(path_out, mode='wb') as f:
+            pq.write_table(table, f)
+        return self.get_path_remote(path_out)
+
+    def _create_folders(self, path: str) -> None:
+        """
+        Create the folders in path if they don't exist.
+        :param path:
+        """
+        pathdir = os.path.dirname(path)
+        if not os.path.exists(pathdir):
+            os.makedirs(pathdir)
